Renamed content type to match naming.
