+++ config/config.js
@@ -1,15 +1,15 @@
 // config.js
 
-var path = require('path'),
-rootPath = path.normalize(__dirname + '/..');
+var path = require("path"),
+  rootPath = path.normalize(__dirname + "/..");
 
 module.exports = {
-    'development': {
-        'database': null,
-        'port': 8000,
-        'installedApps': [
-            'auth',
-            'public'
-        ]
-    }
+  development: {
+    database: null,
+    port: 8000,
+    installedApps: [
+      // 'auth',
+      "public"
+    ]
+  }
 };
